+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The types of changes are:
 * Okta, aws and database credentials can now come from `fidesctl.toml` config [#694](https://github.com/ethyca/fides/pull/694)
 * New `validate` endpoint to test aws and okta credentials [#722](https://github.com/ethyca/fides/pull/722)
 * A new image tagged `ethyca/fidesctl:dev` is published on each push to `main` [781](https://github.com/ethyca/fides/pull/781)
+* A new cli command (`fidesctl sync`) [#765](https://github.com/ethyca/fides/pull/765)
 
 ### Changed